+++ src/main/java/agritech/madaberia/dto/ResponseUserDTO.java
@@ -1,9 +1,12 @@
 package agritech.madaberia.dto;
 
+import agritech.madaberia.domain.Permission;
 import agritech.madaberia.domain.Role;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 
 @AllArgsConstructor
@@ -14,5 +17,8 @@ public class ResponseUserDTO {
     private String firstName;
     private String middleName;
     private String lastName;
+    private String userName;
+    private String password;
     private Role role;
+    private List<Permission> permissionList;
 }
