+++ src/main/java/agritech/madaberia/model/Woreda.java
@@ -1,22 +1,26 @@
-package agritech.madaberia.domain;
+package agritech.madaberia.model;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-public class Kebele {
+public class Woreda {
     @Id
     private Long id;
     private String name;
     private String code;
 
+
     @ManyToOne(optional = false)
-    private Woreda woreda;
+    private Zone zone;
+
+    @OneToMany()
+    private List<Kebele> kebeles;
 }
