+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.datasource.url=jdbc:postgresql://localhost:1234/madaberia-production
 spring.datasource.username=postgres
 spring.datasource.password=root
 
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=update
 #spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 #spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL81Dialect
